+++ api/src/main/java/com/wallet/api/account/infraestructure/AccountRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.repository.CrudRepository;
 
 public interface AccountRepository extends CrudRepository<Account, Integer> {
 
-    Optional<Account> findByDocumentNumber(BigInteger documentNumber);
+    Optional<Account> findByDocumentNumber(int documentNumber);
 }
